requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
chromadb==0.4.22
langchain-community==0.0.13
python-dotenv==1.0.0
rich==13.7.0
textual==0.45.1
typer==0.9.0
pydantic==2.5.3
faiss-cpu==1.7.4
tiktoken==0.7.0
